Lab 3/milestone-3/requirements.txt
torch
torchvision
timm
numpy<2
pillow_heif
matplotlib
onnx>=1.12.0
onnxslim
onnxruntime
apache-beam[gcp]==2.63.0